refactor: move ui components to a seperate folder
