+++ packages/ui-tailwind/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./Block";
-export * from "./Button";
-export * from "./Layout";
-export * from "./Text";
+export * from "./components/Block";
+export * from "./components/Button";
+export * from "./components/Layout";
+export * from "./components/Text";
